requirements.txt
fastapi[standard]
sqlmodel
psycopg2-binary
argon2-cffi
python-jose[cryptography]
fastapi-cors
